hydra: split mpiexec/name_service.c
The name service routines are an independent component. Move the code
into a separate source file.

Use static for internal global and functions.
