+++ src/pm/hydra/mpiexec/Makefile.mk
@@ -10,6 +10,7 @@ mpiexec_hydra_SOURCES = \
     mpiexec/options.c \
     mpiexec/get_parameters.c \
     mpiexec/uiu.c \
+    mpiexec/name_service.c \
     mpiexec/pmiserv_cb.c \
     mpiexec/pmiserv_pmi.c \
     mpiexec/pmiserv_pmi_v1.c \
