+++ src/pm/hydra/mpiexec/pmiserv_pmi.h
@@ -39,22 +39,8 @@ struct HYD_pmcd_pmi_pg_scratch {
     int keyval_dist_count;      /* Number of keyvals distributed */
 };
 
-struct HYD_pmcd_pmi_publish {
-    char *name;
-    char *port;
-    int infokeycount;
-
-    struct HYD_pmcd_pmi_info_keys {
-        char *key;
-        char *val;
-    } *info_keys;
-
-    struct HYD_pmcd_pmi_publish *next;
-};
-
 struct HYD_proxy *HYD_pmcd_pmi_find_proxy(int fd);
 HYD_status HYD_pmcd_pmi_finalize(void);
-HYD_status HYD_pmcd_pmi_free_publish(struct HYD_pmcd_pmi_publish *publish);
 HYD_status HYD_pmcd_pmi_publish(char *name, char *port, int *success);
 HYD_status HYD_pmcd_pmi_unpublish(char *name, int *success);
 HYD_status HYD_pmcd_pmi_lookup(char *name, char **value);
@@ -65,6 +51,5 @@ struct HYD_pmcd_pmi_handle {
 };
 
 extern struct HYD_pmcd_pmi_handle *HYD_pmcd_pmi_handle;
-extern struct HYD_pmcd_pmi_publish *HYD_pmcd_pmi_publish_list;
 
 #endif /* PMISERV_PMI_H_INCLUDED */
